requirements.txt
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
sentence-transformers>=2.3.0
imagehash>=4.3.1
pydantic>=2.5.0
requests>=2.31.0
simhash>=2.1.2
python-dotenv>=1.0.0
